requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
aioresponses>=0.7.4